re-organize GitHub tools
